+++ react/wbooks/src/app/screens/SignUp/actions.js
@@ -1,13 +1,12 @@
-import { createUser } from '../../../services/AuthServices';
-import { normalizeCreateUser } from '../../utils/AuthUtils';
-
-export const actionCreateUser = async user => {
-  const userNormalize = normalizeCreateUser(user);
-  const userCreated = await createUser(userNormalize);
-  if (userCreated.ok) {
-    return { ok: true, data: userCreated.data };
-  }
-  return { ok: false, problem: userCreated.problem };
+const actions = {
+  createUser: 'createUser'
 };
 
-export const setToken = token => localStorage.setItem('token_id', token);
+const actionCreators = {
+  createUser: newUser => ({
+    type: actions.createUser,
+    payload: newUser
+  })
+};
+
+export default actionCreators;
