+++ react/wbooks/src/app/screens/SignUp/reducer.js
@@ -1,0 +1,10 @@
+export const initialState = { user: {} };
+
+export default function reducer(state, action) {
+  switch (action.type) {
+    case 'createUser':
+      return { user: action.payload };
+    default:
+      return state;
+  }
+}
